Updated remove_sitealizer so it doesn't rake abort if sitealizer table is not present.
git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/trunk@1722 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ db/migrate/063_remove_sitealizer.rb
@@ -1,6 +1,11 @@
 class RemoveSitealizer < ActiveRecord::Migration
+  def self.table_exists?(name)
+    ActiveRecord::Base.connection.tables.include?(name)
+  
+  end
+
   def self.up
-    drop_table :sitealizer
+    drop_table :sitealizer if self.table_exists?("sitealizer")
   end
 
   def self.down
